requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
rank_bm25==0.2.2
python-multipart==0.0.6
huggingface-hub<0.19.0
requests
